+++ app/src/main/java/com/sopt/now/ui/login/LoginState.kt
@@ -1,6 +1,7 @@
 package com.sopt.now.ui.login
 
-data class LoginState (
+data class LoginState(
     val isSuccess: Boolean,
+    val userId: String?,
     val message: String
 )
